requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
openai>=1.12.0,<2.0.0
pinecone-client>=3.0.0,<5.0.0
pydantic>=2.0.0,<3.0.0
PyPDF2>=3.0.0,<4.0.0
firebase-admin>=6.0.0,<7.0.0
openpyxl>=3.1.0,<4.0.0
